client/requirements.txt
altgraph==0.17.3
dataclasses==0.8
fbs==1.2.1
lxml==4.9.3
macholib==1.16.2
pefile==2023.2.7
PyInstaller==3.4
PyPDF2==3.0.1
PySide2==5.15.2.1
python-docx==0.8.11
pywin32-ctypes==0.2.2
shiboken2==5.15.2.1
typing_extensions==4.1.1
